Improve robustness when DNS is unreliable
Since DNS resolution on AREDN can sometimes be spotty, resolve the hostname ourselves and cache the IP
address. If DNS lookup for a node fails when it's time to update, fall back to that cached IP address so we
stay in contact with the node. Also expose a sensor to show that resolved IP address.

On my instance, in which AREDN is on a separate VLAN, I've had issues with the systemd-resolved proxy getting
confused and failing hostname lookups for AREDN nodes, resulting in that node appearing unavailable in this
integration. Hopefully this will help, as the network is actually still reachable.